+++ test/test/pytools/viz/dendrogram/test_dendrogram.py
@@ -63,18 +63,53 @@ def test_dendrogram_drawer_text(linkage_matrix: np.ndarray) -> None:
         report_str = str(out.getvalue())
         log.debug(f"\n{report_str}")
 
+        print(report_str)
     report_str_expected = (
         r"""
 ===================================== Test =====================================
 
-G               19% \_____________
-F               17% /             \--------------------------------------------\
-B                6% --------------/                                            |
-H               22% \_____________                                             |
-C                8% /             \______________                              |
-E               14% --------------/              \-----------------------------/
-A                3% --------------/              |
-D               11% -----------------------------/
+G 19% \_________________
+F 17% /                 \------------------------------------------------------\
+B  6% ------------------/                                                      |
+H 22% \_________________                                                       |
+C  8% /                 \_________________                                     |
+E 14% ------------------/                 \------------------------------------/
+A  3% ------------------/                 |
+D 11% ------------------------------------/
+
+"""
+    )[1:]
+    assert report_str == report_str_expected
+
+    linkage_tree = LinkageTree(
+        scipy_linkage_matrix=linkage_matrix,
+        leaf_names=leaf_names,
+        leaf_weights=[(w + 1) / 36 for w in range(8)],
+        max_distance=5.0,
+        distance_label="distance",
+        leaf_label="label",
+        weight_label="weight",
+    )
+
+    with StringIO() as out:
+        dd = DendrogramDrawer(style=DendrogramReportStyle(out=out))
+        dd.draw(data=linkage_tree.sort_by_weight(), title="Test")
+        report_str = str(out.getvalue())
+        log.debug(f"\n{report_str}")
+
+        print(report_str)
+    report_str_expected = (
+        r"""
+===================================== Test =====================================
+
+G 19% \_____________
+F 17% /             \--------------------------------------------\
+B  6% --------------/                                            |
+H 22% \_____________                                             |_____________
+C  8% /             \______________                              |
+E 14% --------------/              \-----------------------------/
+A  3% --------------/              |
+D 11% -----------------------------/
 
 """
     )[1:]